backend/pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitagpt-backend"
version = "0.1.0"
description = "FastAPI RAG backend for GitaGPT"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "numpy>=1.26.0",
  "PyMuPDF>=1.24.0"
]

[project.optional-dependencies]
rag = [
  "faiss-cpu>=1.8.0",
  "sentence-transformers>=3.0.0"
]
openai = [
  "openai>=1.40.0"
]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100

[tool.setuptools.packages.find]
include = ["app*"]
